Samadhan Kendra2/Samadhan Kendra2/Backend/requirement.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.2
django-cors-headers>=4.4
psycopg2-binary>=2.9
dj-database-url>=2.2
python-dotenv>=1.0